pyproject.toml
[tool.ruff]
line-length = 88
target-version = "py38"
src = ["ragflow"]
exclude = [".git", ".venv", "__pycache__", ".mypy_cache"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "ERA",  # eradicate
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "D",    # pydocstyle
    "S",    # flake8-bandit
]
ignore = [
    "D203",  # one-blank-line-before-class
    "D212",  # multi-line-summary-first-line
    "E501",  # line-too-long (handled by formatter)
    "B904",  # use raise from
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S", "D"]
"examples/*" = ["D", "E501", "ERA001"]
"**/examples/*.py" = ["D", "E501", "ERA001"]

[tool.bandit]
exclude_dirs = ["tests", "examples"]
skips = ["B101"]

[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["ragflow*"]

[project]
name = "ragflow"
version = "0.1.0"
description = "A high-level framework for Retrieval Augmented Generation"
authors = [
    {name = "RAGFlow Team"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.3.25",
    "sentence-transformers>=2.2.2",
    "pysqlite3-binary>=0.4.4",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.3.3",
    "ruff>=0.0.282",
    "bandit>=1.7.5",
]
test = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
]
docs = [
    "mkdocs>=1.4.3",
    "mkdocs-material>=9.1.17",
]
